Move header drag and drop actions to its own files
