+++ packages/swayze/lib/src/core/intents/intents.dart
@@ -1,3 +1,4 @@
+export 'drag_n_drop_intents.dart';
 export 'inline_editor_intents.dart';
 export 'selection_intents.dart';
 export 'swayze_intent.dart';
