+++ packages/swayze/lib/src/core/intents/selection_intents.dart
@@ -88,46 +88,3 @@ class HeaderSelectionUpdateIntent extends SwayzeIntent {
     required this.axis,
   });
 }
-
-// TODO: [victor] doc
-class HeaderDragStartIntent extends SwayzeIntent {
-  final Range headers;
-  final Axis axis;
-  final Offset draggingPosition;
-
-  const HeaderDragStartIntent({
-    required this.headers,
-    required this.axis,
-    required this.draggingPosition,
-  });
-}
-
-// TODO: [victor] doc
-class HeaderDragUpdateIntent extends SwayzeIntent {
-  final int header;
-  final Axis axis;
-  final Offset draggingPosition;
-
-  const HeaderDragUpdateIntent({
-    required this.header,
-    required this.axis,
-    required this.draggingPosition,
-  });
-}
-
-// TODO: [victor] doc
-class HeaderDragEndIntent extends SwayzeIntent {
-  final int header;
-  final Axis axis;
-
-  const HeaderDragEndIntent({
-    required this.header,
-    required this.axis,
-  });
-}
-
-class HeaderDragCancelIntent extends SwayzeIntent {
-  final Axis axis;
-
-  const HeaderDragCancelIntent(this.axis);
-}
