+++ packages/swayze/lib/src/widgets/headers/gestures/header_gesture_detector.dart
@@ -7,6 +7,7 @@ import 'package:swayze_math/swayze_math.dart';
 
 import '../../../../controller.dart';
 import '../../../../intents.dart';
+import '../../../core/intents/drag_n_drop_intents.dart';
 import '../../../core/viewport_context/viewport_context.dart';
 import '../../../core/viewport_context/viewport_context_provider.dart';
 import '../../../helpers/scroll/auto_scroll.dart';
